+++ tests/Test/Bridge/Doctrine/Type/DoctrineEnumListTypeTest.php
@@ -45,7 +45,7 @@ class DoctrineEnumListTypeTest extends TestCase
         $this->assertSame([], $this->type->convertToPHPValue('', $this->platform));
         $this->assertEquals(
             Enum::fromValues([Enum::VALID_VALUE]),
-            $this->type->convertToPHPValue(serialize([Enum::VALID_VALUE]), $this->platform)
+            $this->type->convertToPHPValue(serialize([Enum::VALID_VALUE]), $this->platform),
         );
     }
 
@@ -92,7 +92,7 @@ class DoctrineEnumListTypeTest extends TestCase
         $this->assertSame(serialize([]), $this->type->convertToDatabaseValue([], $this->platform));
         $this->assertEquals(
             serialize(['value1', 'value2']),
-            $this->type->convertToDatabaseValue([Enum::fromValue('value1'), Enum::fromValue('value2')], $this->platform)
+            $this->type->convertToDatabaseValue([Enum::fromValue('value1'), Enum::fromValue('value2')], $this->platform),
         );
     }